move chatroom code to chatrooms module
